+++ src/components/index.js
@@ -1,6 +1,7 @@
 export { default as App } from './App/App';
 export { default as Auth } from './Auth/Auth';
 export { default as Avatar } from './Avatar/Avatar';
+export { default as AvatarImg } from './Avatar/AvatarImg';
 export { default as Board } from './Board/Board';
 export { default as Button } from './Button/Button';
 export { default as InputArea } from './InputArea/InputArea';